config: 设置EEPROM和Rolling Code模式并保护校准区
1. 配置Keil Code区为EEPROM+Rolling Code模式
2. 明确校准数据保存区为0x3F00~0x3F7F
3. 0x3F80以后当前未使用，程序不主动擦写
4. 增加Flash擦写边界保护，防止误写校准区外数据
5. 保持业务逻辑、通信协议和校准流程不变

+++ YAT-60V3A-Uart-1-wire/App/cal.c
@@ -156,7 +156,7 @@ void save_cal_data(void)
 {
     cal_val.flag = CAL_FLAG_V2;
     flash_ferase_sector(FLASH_CAL_ADDR);
-    flash_program(FLASH_CAL_ADDR, (u8 *)&cal_val, (u16)sizeof(cal_val));
+    flash_program(FLASH_CAL_ADDR, (u8 *)&cal_val, (u8)sizeof(cal_val));
 }
 
 void read_cal_data(void)
